requirements.txt
openai>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
httpx>=0.27.0
dhanhq>=1.1.0
